Fix: MR_PSYCHIC was not applied correctly for all monsters with the flag, other Illithid tweaks. Needed to refactor resistances to be able to handle the extra inclusion.
Also in this commit - Illithids do double damage with their psionic attack at experience level 26 or higher. Illithids wearing a helm made of any 'heavy' metal (basically any metallic helmet minus mithril) will have its innate psionic attack blocked.

+++ include/monst.h
@@ -88,7 +88,7 @@ struct monst {
     uchar m_ap_type;      /* what mappearance is describing, m_ap_types */
 
     schar mtame;                /* level of tameness, implies peaceful */
-    unsigned short mextrinsics; /* low 8 correspond to mresists */
+    unsigned long mextrinsics;  /* low 9 correspond to mresists */
     int mspec_used;             /* monster's special ability attack timeout */
 
 #define M_SEEN_NOTHING 0x0000
